+++ code/graphics/grstub.cpp
@@ -254,7 +254,7 @@ void gr_stub_set_buffer(int idx)
 {
 }
 
-void gr_stub_update_buffer_object(int handle, uint size, void* data)
+void gr_stub_update_buffer_object(int handle, size_t size, void* data)
 {
 
 }
@@ -274,7 +274,7 @@ int gr_stub_create_stream_buffer()
 	return -1;
 }
 
-void gr_stub_render_stream_buffer(int buffer_handle, int offset, int n_verts, int flags)
+void gr_stub_render_stream_buffer(int buffer_handle, size_t offset, size_t n_verts, int flags)
 {
 }
